refactor: move transposeNotes to Song
